requirement.txt
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pycodestyle==2.5.0
requests==2.22.0
todoist-python==8.1.1
urllib3==1.25.7
